Things check description
Added clickCollider - collider over entire map that checks for clicks everywhere but ui and other towers. Is disabled when tower is being placed.
towerButton.cs added to hide towerui if a tower is selected and u drag tower from bank.
Also shows range indicator if tower is being dragged in line 183
Added name field to towers and sell amount field.
Added sellTower method and button to towerUI
Added exit button to towerUI
Added detectClick script to all levelColliders
On intialization of a tower the indicator is set to default range and will be updated accordingly in showTowerUI method.